+++ require.php
@@ -7,3 +7,4 @@ require_once path_join(FRONTEND_DIR, 'library.php');
 require_once path_join(LIBS_DIR, 'file.php');
 require_once path_join(LIBS_DIR, 'log.php');
 require_once path_join(BUILDER_DIR, 'minimize.php');
+require_once path_join(BUILDER_DIR, 'asset.php');